refactor: create useDeployment hook
