+++ common/model_version.ts
@@ -15,3 +15,21 @@ export enum MODEL_VERSION_STATE {
   deployFailed = 'DEPLOY_FAILED',
   registerFailed = 'REGISTER_FAILED',
 }
+
+export const isModelDeployable = (state: MODEL_VERSION_STATE) => {
+  if (
+    state === MODEL_VERSION_STATE.undeployed ||
+    state === MODEL_VERSION_STATE.registered ||
+    state === MODEL_VERSION_STATE.deployFailed
+  ) {
+    return true;
+  }
+  return false;
+};
+
+export const isModelUndeployable = (state: MODEL_VERSION_STATE) => {
+  if (state === MODEL_VERSION_STATE.deployed || state === MODEL_VERSION_STATE.partiallyDeployed) {
+    return true;
+  }
+  return false;
+};
